Code to compute normal (problem)
I have written the code to compute (using a brute-force approach) the normal of each edge corresponding to each element, assuming that the function getNodes was able to compute the coordinates of each node. In fact, it can only return the coordinate of each point (defined in the file .geo), such that I have put everything linked to that part in comment. Also, the getNormal command seems to be designed for 3D cases, since it  shuld compute normals to surfaces => search in examples/sdk or ask next week