Tests to understand how the Basis functions & Jacobians work
I've added a function that retrieves using gmsh the basis functions & the jacobians, to understand how it works and how data is handled. The file is obvisoulsy very temporary, and will be deleted when we will have properly understood everything.
N.B.: maybe [det J] is independent of the GP, due to the form of the application between the reference and physical frame (e.g. isoparametric, but is this property general ?). It would mean that the [M] matrix could be build without caring about the "orientation" of the element. Nonetheless, I doubt that this property holds in general (but let's hope :smile: !). Furthemore, we still need to know which elements are where for the computation of the normals (i.e. the [S] matrix), as well as for the flux.