Normal computation comments
I have reviewed @imperators79 contribution to the normal computation. Evrything seems ok (and in particular I have not modified anything); I have just added comments.

Nonetheless, I have 3 remarks:
(i) The current code does not handle the case where we have more than 2 nodes per element (and for high order elements we defenitely will get more than 2 nodes !). I don't know if taking that into account is possible for us at the current moment.
(ii) The efficiency could be improved, by considering that almost every normal is computed twice (adjacent elements) => maybe we could think of something to take that into account
(iii) The problem with the clockwise/counter-clockwise may be solved by considering the 3rd component of the jacobian of the element, i.e. wheter or not the element was flipped. Maybe then, there is a rule for the order in which the nodes per edge are organized.