[M] computation with new structure
- update of the buildM function to be functional with the new strucutre
- precompute the components w_k*l_i*l_j when a type of element is added (i.e., this is done in Mesh2D !)
- add a function that computes the number of unknwons in Mesh2D => allows to define the [M] matrix
- add \t for the tabulations in the display function

N.B.: Apparently, when the mesh is specified in the arguments of the function as a const, I can't access one of its fields which is a map (more precisely I want to get something like mesh.elementProperties2D[element.elementType2D]). There may be a solution but I haven't found it yet.