Check & correction of [M], [Sx], [Sy]
I have manually checked that the DG matrices were correct.
- For [M], there was no mistake
- For [Sx]/[Sy], there were three mistakes:
(i) when considering the fast inverse matrix computation (in the 2*2 matrix), we were forgetting the fact that dz/dzeta might be different from 1 (in fact, it is probably equal to -1). Therefore, we have to multiply the dxi/dx, ... by "sign", where "sign" corresponds to this component
(ii) I have no idea why, but we were writting dlidxi and so on, instead of dljdxi and so on. Since [Sx]/[Sy] are expected not to be symmetric, this correction is important
(iii) For this fact inverse calculation, we should not forget that if dx1/dX2 is at position (2, 1), then dX2/dx1 will be at position (1, 2), and not (2, 1), in the inverse matrix. This is what was done before.

Remark: to compute manually the values, I used this link: http://math2.uncc.edu/~shaodeng/TEACHING/math5172/Lectures/Lect_15.PDF. Something that was bothering me is that in fact the weights (in this source as well as our FEM course) are "normalized" is some way, meaning that for 1GP, we get a weight of 1, even tough the area of the triangle is 1/2... In gmsh, this 1/2 is already taken into account.