Explanation [S] matrices + comments (cont'd)
Changes:
- begin the code readibility improvements (folders for functions, indentation and comments)
- add the "inverted" rectangle for testing purposes

Explanation for the [S] matrices issue:
I've looked at my calculations and compared them with the theory that is described in the reference book. In fact, we defined our matrices [S] as the transpose of the matrices [S] that are defined in the book. This mistake comes from the fact that the unknown vector is represented by its values u_j (and not u_i !), to compute the matrix-vector product with a matrix of the form [M]_{ij}. In any case, the solution was simply to redefine in buildS the way [S] is constructed (it just comes to computing its transpose, that is, to swipe the i and j indices). In what you had done "experimentally" to try to find a solution @imperators79, you inverted the strong and weak forms (but it does not change anything; both work). I've put in red in the first overleaf report the correct formulation of the weak and strong forms.